Implement PDLA_Anyval datatype as union to support 64bit indexing
This adds a union datatype to add 64bit integer support to
the original PDLA-2.x types which assumed that double was
capable of holding all the "lesser" types.  With the PDLA_Anyval
type, we can correctly handle 64bit integer datatypes and
avoid errors and loss of precision due to conversions to or
from PDLA_Double.  Thanks also to kmx for his contributions
to finish the implementation.