Clarify documentation in PDLA::Fit::LM. Add small test-suite case.
Mostly this clarifies the usage of the 'sig' argument to lmfit.
'sig' was short for 'sigma', I have changed the former to the latter.
Confusion sowed because in the example the variable was called
$wt ("weight"), when in reality the weights turn out to be 1/(sigma**2).

I added a small test suite in t/lm.t.  This uses the data 'silver.dat'
from Gnuplot, which I believe to be compatible with PDLA's license.